scripts/requirements.txt
google-genai
google-generativeai
deep-translator
tqdm
Pillow